Feat/unit test

+++ src/interfaces/IGovernance.sol
@@ -14,7 +14,7 @@ enum ProposalState {
 }
 
 interface IGovernance {
-    function state(uint proposalId) external view returns (ProposalState);
-    function proposer(uint proposalId) external view returns (address);
+    function state(uint256 proposalId) external view returns (ProposalState);
+    function proposer(uint256 proposalId) external view returns (address);
     function proposalCount() external view returns (uint256);
 }
